refactor: mongbi 관련 말풍선 내용, 이미지 경로를 core/constants로 이동
